Añadiendo diagrama pdf

+++ src/data/DiagramaDB/DATABASEXLR8-CREACIONTABLAS.sql
@@ -39,10 +39,10 @@ CREATE TABLE IF NOT EXISTS `xrl8`.`products` (
   `color` VARCHAR(100) NOT NULL,
   `genre_product` VARCHAR(100) NOT NULL,
   `type` VARCHAR(100) NOT NULL,
-  `new` TINYINT(1) NOT NULL,
+  `new` INT(11) NOT NULL,
   `createdAt` DATE NOT NULL,
   `updatedAt` DATE NOT NULL,
-  `deletedAt` TINYINT(1) NOT NULL,
+  `deletedAt` INT(11) NOT NULL,
   PRIMARY KEY (`id`))
 ENGINE = InnoDB
 AUTO_INCREMENT = 4
